pyproject.toml
[project]
name = "ragapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=25.9.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.15",
    "langchain-openai>=0.3.35",
    "langchain-pinecone>=0.2.12",
    "langchain-text-splitters>=0.3.5",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.2.1",
]
